Reworking GetFeatures API
