+++ GetFeatures/utils/geo.py
@@ -7,6 +7,11 @@ import numpy as np
 from rasterio.sample import sample_gen
 from utils.cache import memoize_geospatial_with_persistence
 
+import pyproj
+from shapely.geometry import Point
+from shapely.ops import transform
+
+
 # @memoize_geospatial_with_persistence('/tmp/xr_vectorize_cache.pkl')
 def xr_vectorize(
     da,
@@ -131,10 +136,29 @@ def add_geobox(ds, crs=None):
 
 
 # @memoize_geospatial_with_persistence('/tmp/extract_points.pkl')
-def extract_points(ds, gdf, column_name) -> gpd.GeoDataFrame:
+def extract_z_values(ds, gdf, column_name) -> gpd.GeoDataFrame:
     # note the extra 'z' dimension that our results will be organized along
     da_x = xr.DataArray(gdf.geometry.x.values, dims=['z'])
     da_y = xr.DataArray(gdf.geometry.y.values, dims=['z'])
-    results = ds.isel(band=0).sel(x=da_x, y=da_y, method='nearest')
+    results = ds.sel(x=da_x, y=da_y, method='nearest')
     gdf[column_name] = results.values
     return gdf
+
+# Convert GeoJSON to GeoDataFrame
+def geojson_to_geodataframe(geojson):
+    features = geojson["features"]
+    geometries = [shape(feature["geometry"]) for feature in features]
+    properties = [feature["properties"] for feature in features]
+
+    gdf = gpd.GeoDataFrame(properties, geometry=geometries)
+    return gdf
+
+
+def transform_point(x, y, crs):
+    pt = Point(x, y)
+
+    init_crs = pyproj.CRS(crs)
+    wgs84 = pyproj.CRS('EPSG:4326')
+
+    project = pyproj.Transformer.from_crs(init_crs, wgs84, always_xy=True).transform
+    return transform(project, pt)
